Add initial meeting message type enum values. (#40)
To support #38 , add new enumeration types corresponding to Exchange
meeting messages. This commit also adds a convenience method to obtain a
reference to the internally contained Message object. This enables
reduction of boilerplate in match statements at the point of use where
the primary purpose is to obtain the contained message.