implement `Deserialize` for `MessageXml`
This is an extensive change to how `MessageXml` works, and loses the ability to
store the raw XML within the type (most attributes in particular will now be
implicitly stripped, as will new fields in known types). This change allows
`MessageXml` as a member in other types that derive `Deserialize`.