implement `Deserialize` for `MessageXml` (#52)
This is an extensive change to how `MessageXml` works, and loses the
ability to store the raw XML within the type (most attributes in
particular will now be implicitly stripped, as will new fields in known
types). This change allows `MessageXml` as a member in other types that
derive `Deserialize`.

First half of addressing [Bug
1971330](https://bugzilla.mozilla.org/show_bug.cgi?id=1971330).