Add CalendarItem item type. (#71)
I know we've never seen one of these in the wild, but going through our
remaining items last week, I got a little nervous that this was still an
outstanding item. Since it's small, I figured we could try and get it
into 145 if we can get it through on time.